Rename subpackage name by signals
