+++ pandas_ta/core.py
@@ -1225,10 +1225,10 @@ class AnalysisIndicators(BasePandasObject):
 
 
     # Events indicators
-    def rsi_event(self, close=None, above_val=None, below_val=None, length=None, drift=None, offset=None, **kwargs):
+    def rsi_signals(self, close=None, above_val=None, below_val=None, length=None, drift=None, offset=None, **kwargs):
         close = self._get_column(close, 'close')
-        from pandas_ta.event.rsi_event import rsi_event
-        result = rsi_event(close=close, above_val=above_val, below_val=below_val, length=length, drift=drift, offset=offset, **kwargs)
+        from pandas_ta.signals.rsi_signals import rsi_signals
+        result = rsi_signals(close=close, above_val=above_val, below_val=below_val, length=length, drift=drift, offset=offset, **kwargs)
         self._add_prefix_suffix(result, **kwargs)
         self._append(result, **kwargs)
         return result
